requirements.txt
aiogram==3.13.1
aiohttp==3.10.11
aiosqlite==0.20.0
python-dotenv==1.0.1
Pillow==11.0.0
httpx==0.27.2
